Replace array_ref by gsl::span.

+++ inc/goldfish/sax_writer.h
@@ -137,7 +137,7 @@ namespace goldfish { namespace sax
 		auto write(const std::string_view& text)
 		{
 			auto stream = start_string(text.size());
-			stream.write_buffer({ reinterpret_cast<const byte*>(text.data()), text.size() });
+			stream.write_buffer({ reinterpret_cast<const byte*>(text.data()), gsl::narrow_cast<std::ptrdiff_t>(text.size()) });
 			return stream.flush();
 		}
 
@@ -195,7 +195,7 @@ namespace goldfish { namespace sax
 			{
 				// We read the entire stream
 				auto output_stream = create_writer_with_size(cb);
-				output_stream.write_buffer({ buffer, cb });
+				output_stream.write_buffer({ buffer, gsl::narrow_cast<std::ptrdiff_t>(cb) });
 				return output_stream.flush();
 			}
 			else
